move partition/test_client.py to test/test_partition.py
