+++ lib/arynsdk/arynsdk/partition/__init__.py
@@ -1,3 +1,3 @@
-from .client import partition_file
+from .partition import partition_file
 
 __all__ = ["partition_file"]
